// New/shared-lib/services/qualtrics/constants/qualtrics-properties.enum.ts
export enum QualtricsProperty {
  PAGE_NAME = 'Page_Name',
  PARTY_ID = 'PartyId',
  PLAN = 'PlanId',
  EMAIL = 'Email',
  PHONE = 'Phone',
  FIRST_NAME = 'First_Name',
  LAST_NAME = 'Last_Name',
  FIRST_TIME_USER = 'First_Time_User',
  APP_ID = 'AppId',
  CLIENT_ID = 'ClientId',
  FEEDBACK = 'feedback',
  CONTACT_VOYA_SUPPORT = 'contact_voya_support',
  DEVICE_TYPE = 'deviceType',
  MYVOYAGE_ENABLED = 'myVoyageEnabled',
}